+++ CHANGELOG.md
@@ -7,5 +7,6 @@
 - Add get() method to Account
 - Add create() method to Account
 - Add update() method to Account
+- Add delete() method to Account
 - Add proper testing for ToshlException
 - Improve string representation of ToshlException